+++ LogsManager/Classes/_Message Loggers/_Base/BaseAbstractTextLogger.swift
@@ -1,5 +1,5 @@
 //
-//  BaseAbstractLogger.swift
+//  BaseAbstractTextLogger.swift
 //  LogsManager
 //
 //  Created by Anton Plebanovich on 3/2/18.
@@ -11,11 +11,11 @@ import Foundation
 import CocoaLumberjack
 
 
-/// Abstract base class inherited from DDAbstractLogger.
+/// Abstract base class inherited from DDAbstractLogger that represents text logger.
 /// Override process(message:) in child classes
-open class BaseAbstractLogger: DDAbstractLogger, BaseTextLogger {
+open class BaseAbstractTextLogger: DDAbstractLogger, BaseLogger {
     
-    // ******************************* MARK: - BaseTextLogger
+    // ******************************* MARK: - BaseLogger
     
     public let logLevel: DDLogLevel
     public let mode: LoggerMode
@@ -38,17 +38,20 @@ open class BaseAbstractLogger: DDAbstractLogger, BaseTextLogger {
     override open func log(message logMessage: DDLogMessage) {
         guard shouldLog(message: logMessage) else { return }
         
-        let formattedMessage: String
+        var formattedMessage: String
         if let logFormatter = value(forKey: "_logFormatter") as? DDLogFormatter {
             formattedMessage = logFormatter.format(message: logMessage) ?? logMessage.message
         } else {
             formattedMessage = logMessage.message
         }
         
+        if let normalizedError = logMessage.parameters?.normalizedError { formattedMessage.append("\n\(normalizedError)") }
+        if let normalizedData = logMessage.parameters?.normalizedData { formattedMessage.append("\n\(normalizedData)") }
+        
         process(message: logMessage, formattedMessage: formattedMessage)
     }
     
-    // ******************************* MARK: - Public Methods
+    // ******************************* MARK: - Open Methods
     
     /// Abstract method that should be overridden. Message validation already done. You shouldn't call super in your implementation.
     open func process(message logMessage: DDLogMessage, formattedMessage: String) {